+++ src/components/ProductInfo/ProductInfo.module.scss
